ASC-127 Moved parameter setup to core folder
